backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
aiofiles==24.1.0
pydantic==2.10.0
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
redis==5.2.0
celery==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.28.0
pandas==2.2.3
numpy==2.2.0
scikit-learn==1.5.2
langchain==0.3.12
langchain-community==0.3.12
ollama==0.4.5
python-docx==1.1.2
openpyxl==3.1.5
reportlab==4.2.5
pypdf2==3.0.1
pyyaml==6.0.2
python-dateutil==2.9.0
regex==2024.11.6
pytz==2024.2
aiohttp==3.11.11
beautifulsoup4==4.12.3
lxml==5.3.0
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
flake8==7.1.1
mypy==1.14.0